Refactor pipeline setup command
