+++ commands/pipelines/setup/getGitHubToken.js
@@ -1,0 +1,9 @@
+function getGitHubToken (kolkrabbi) {
+  return kolkrabbi.getAccount().then((account) => {
+    return account.github.token
+  }, () => {
+    throw new Error('Account not connected to GitHub.')
+  })
+}
+
+module.exports = getGitHubToken
