+++ commands/pipelines/setup/hasCIFlag.js
@@ -1,0 +1,13 @@
+const api = require('../../../lib/api')
+
+function* hasCIFlag (heroku) {
+  let hasFlag
+  try {
+    hasFlag = (yield api.getAccountFeature(heroku, 'ci')).enabled
+  } catch (error) {
+    hasFlag = false
+  }
+  return hasFlag
+}
+
+module.exports = hasCIFlag
